requirements.txt
asgiref==3.2.10
Django==3.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
joblib==0.16.0
numpy==1.19.0
pandas==1.0.5
plotly==4.8.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
